+++ Libs/Project/CMakeLists.txt
@@ -1,14 +1,14 @@
 set(SOURCES
   Project.cpp
   Subject.cpp
-  Settings.cpp
+  Parameters.cpp
   Variant.cpp
   )
 
 set(HEADERS
   Project.h
   Subject.h
-  Settings.h
+  Parameters.h
   Variant.h
   )
